chore(connect): move logs to @trezor/utils
